backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
supabase==2.9.1
httpx==0.27.2
langchain==0.3.7
langchain-google-genai==2.0.4
langchain-community==0.3.7
langchain-core==0.3.15
google-generativeai==0.8.3
pdfplumber==0.11.4
python-jose[cryptography]==3.3.0
tenacity==9.0.0
python-dotenv==1.0.1
